backend/requirements.txt
# Lightweight requirements for Render free tier (512MB RAM limit)
# Removes: torch, transformers, sentence-transformers, cuda packages

# Core FastAPI
fastapi==0.111.0
uvicorn[standard]==0.42.0
python-dotenv==1.2.2
python-multipart==0.0.22

# PDF Processing
pypdf==6.9.1

# LLM Integration (Groq only — no local models)
langchain==1.2.12
langchain-core==1.2.19
langchain-groq==1.1.2
groq==0.37.1

# Vector Database (Pinecone)
langchain-pinecone==0.2.13
pinecone==6.0.0

# File Storage (Supabase)
supabase==2.28.2

# Language Detection (lightweight)
langdetect==1.0.9

# Utilities
requests==2.32.5
pydantic==2.12.5
pydantic-settings==2.13.1
